+++ ChangeLog.md
@@ -7,6 +7,7 @@
     Release date: 
 
     - Define 'bodyHeight' set() parameter to handle dynamic height contents
+    - Add NEW button
 
 ### 1.9.0